+++ google2/src/main/scala/org/broadinstitute/dsde/workbench/google2/GoogleDataprocService.scala
@@ -1,20 +1,20 @@
 package org.broadinstitute.dsde.workbench.google2
 
+import ca.mrvisser.sealerate
+import cats.Parallel
 import cats.effect._
 import cats.effect.concurrent.Semaphore
 import cats.mtl.Ask
 import com.google.api.gax.core.FixedCredentialsProvider
 import com.google.api.services.compute.ComputeScopes
 import com.google.auth.oauth2.GoogleCredentials
+import com.google.cloud.compute.v1.Operation
 import com.google.cloud.dataproc.v1.{RegionName => _, _}
 import io.chrisdavenport.log4cats.StructuredLogger
 import org.broadinstitute.dsde.workbench.RetryConfig
 import org.broadinstitute.dsde.workbench.google2.util.RetryPredicates
 import org.broadinstitute.dsde.workbench.model.TraceId
 import org.broadinstitute.dsde.workbench.model.google.{GcsBucketName, GoogleProject}
-import ca.mrvisser.sealerate
-import cats.Parallel
-import com.google.cloud.compute.v1.Operation
 
 import scala.collection.JavaConverters._
 
@@ -78,19 +78,12 @@ object GoogleDataprocService {
     pathToCredential: String,
     blocker: Blocker,
     blockerBound: Semaphore[F],
-    regionName: RegionName,
-    retryConfig: RetryConfig = RetryPredicates.standardRetryConfig
+    regionName: RegionName
   ): Resource[F, GoogleDataprocService[F]] =
     for {
       credential <- credentialResource(pathToCredential)
       scopedCredential = credential.createScoped(Seq(ComputeScopes.CLOUD_PLATFORM).asJava)
-      interpreter <- fromCredential(googleComputeService,
-                                    scopedCredential,
-                                    blocker,
-                                    regionName,
-                                    blockerBound,
-                                    retryConfig
-      )
+      interpreter <- fromCredential(googleComputeService, scopedCredential, blocker, regionName, blockerBound)
     } yield interpreter
 
   def resourceFromUserCredential[F[_]: StructuredLogger: Async: Timer: Parallel: ContextShift](
@@ -98,19 +91,12 @@ object GoogleDataprocService {
     pathToCredential: String,
     blocker: Blocker,
     blockerBound: Semaphore[F],
-    regionName: RegionName,
-    retryConfig: RetryConfig = RetryPredicates.standardRetryConfig
+    regionName: RegionName
   ): Resource[F, GoogleDataprocService[F]] =
     for {
       credential <- userCredentials(pathToCredential)
       scopedCredential = credential.createScoped(Seq(ComputeScopes.CLOUD_PLATFORM).asJava)
-      interpreter <- fromCredential(googleComputeService,
-                                    scopedCredential,
-                                    blocker,
-                                    regionName,
-                                    blockerBound,
-                                    retryConfig
-      )
+      interpreter <- fromCredential(googleComputeService, scopedCredential, blocker, regionName, blockerBound)
     } yield interpreter
 
   def fromCredential[F[_]: StructuredLogger: Async: Timer: Parallel: ContextShift](
@@ -118,8 +104,7 @@ object GoogleDataprocService {
     googleCredentials: GoogleCredentials,
     blocker: Blocker,
     regionName: RegionName,
-    blockerBound: Semaphore[F],
-    retryConfig: RetryConfig
+    blockerBound: Semaphore[F]
   ): Resource[F, GoogleDataprocService[F]] = {
     val settings = ClusterControllerSettings
       .newBuilder()
@@ -129,7 +114,7 @@ object GoogleDataprocService {
 
     for {
       client <- backgroundResourceF(ClusterControllerClient.create(settings))
-    } yield new GoogleDataprocInterpreter[F](client, googleComputeService, blocker, blockerBound, retryConfig)
+    } yield new GoogleDataprocInterpreter[F](client, googleComputeService, blocker, blockerBound)
   }
 }